% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_app}
\alias{run_app}
\title{Run App}
\usage{
run_app(production = FALSE, ...)
}
\arguments{
\item{...}{Additional parameters to pass to \link[quarto:quarto_serve]{quarto::quarto_serve}.}
}
\description{
Run application
}
