% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary.R
\name{summary_output}
\alias{summary_output}
\title{Summary Output Function}
\usage{
summary_output(id, output)
}
\arguments{
\item{id}{Output element ID. One of \code{"summary_stats"} or \code{"summary_table"}.}

\item{output}{Shiny output object.}
}
\description{
Creates output placeholders for the summary module.
}
