% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary.R
\name{summary_server}
\alias{summary_server}
\title{Summary Server Function}
\usage{
summary_server(input, output)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}
}
\description{
Server logic for the summary module. Computes descriptive statistics for the
selected variables.
}
