% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tatami-package.R
\docType{package}
\name{tatami-package}
\alias{tatami}
\alias{tatami-package}
\title{tatami: A Framework for Building Quarto Dashboards}
\description{
An opinionated framework for building modular Quarto dashboards.
}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph@tidy-intelligence.com}

}
\keyword{internal}
